backend/requirements.txt
flask
pandas
flask-cors
py-dateinfer